+++ app/src/main/java/com/mad/assignment11453798/activity/SettingsActivity.java
@@ -1,22 +1,13 @@
-package com.mad.assignment11453798.Activity;
+package com.mad.assignment11453798.activity;
 
-import android.content.res.Resources;
 import android.os.Bundle;
 import android.preference.PreferenceActivity;
 import android.preference.PreferenceFragment;
 
-import com.crashlytics.android.Crashlytics;
-import com.facebook.FacebookSdk;
-import com.facebook.appevents.AppEventsLogger;
 import com.mad.assignment11453798.R;
-import com.twitter.sdk.android.Twitter;
-import com.twitter.sdk.android.core.TwitterAuthConfig;
-
-import io.fabric.sdk.android.Fabric;
 
 /**
  * SettingsActivity Class
- * Gets social media access tokens via login
  */
 public class SettingsActivity extends PreferenceActivity {
