+++ app/src/main/java/com/mad/assignment11453798/adapter/TwitterTweetAdapter.java
@@ -1,4 +1,4 @@
-package com.mad.assignment11453798.Adapter;
+package com.mad.assignment11453798.adapter;
 
 import android.content.Context;
 import android.support.v7.widget.RecyclerView;
@@ -7,7 +7,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
-import com.mad.assignment11453798.Pojo.TwitterTweet;
+import com.mad.assignment11453798.pojo.TwitterTweet;
 import com.mad.assignment11453798.R;
 
 import java.util.List;
