--- app/src/main/res/layout/activity_settings.xml
@@ -1,10 +1,0 @@
-<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_height="match_parent"
-    android:layout_width="match_parent">
-    <PreferenceCategory
-        android:title="Account">
-        <ListPreference
-            android:title="HELLO">
-        </ListPreference>
-    </PreferenceCategory>
-</PreferenceScreen>
