+++ app/src/main/res/menu/menu_main.xml
@@ -3,6 +3,10 @@
     xmlns:tools="http://schemas.android.com/tools"
     tools:context=".MainActivity">
     <item
+        android:id="@+id/menu_accounts"
+        app:showAsAction="always"
+        android:title="Accounts"/>
+    <item
         android:id="@+id/menu_settings"
         app:showAsAction="never"
         android:title="Settings"/>
