+++ app/src/main/res/xml/preferences.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android">
+    <PreferenceCategory
+        android:title="Accounts">
+        <Preference
+            android:title="Testing what is">
+        </Preference>
+        <RingtonePreference
+            android:title="Ringtone Preference"/>
+    </PreferenceCategory>
+    <EditTextPreference android:title="Your Name"
+        android:key="username"
+        android:summary="Please provide your username"/>
+    <CheckBoxPreference android:title="Application Updates"
+        android:defaultValue="false"
+        android:summary="This option if selected will allow the application to check for latest versions."
+        android:key="applicationUpdates" />
+    <ListPreference     android:title="Download Details"
+        android:summary="Select the kind of data that you would like to download"
+        android:key="downloadType"
+        android:defaultValue="1"
+        android:entries="@array/listArray"
+        android:entryValues="@array/listValues" />
+    <!--<PreferenceCategory-->
+    <!--android:title="Account">-->
+    <!--<ListPreference-->
+    <!--android:title="HELLO">-->
+    <!--</ListPreference>-->
+    <!--</PreferenceCategory>-->
+    <!--</PreferenceScreen>-->
+</PreferenceScreen>
